University optimisation project
In this project, we optimized house floor designs to maximize the profit of a hypothetical developer. All conditions, possible combinations of designs for each floor, as well as additional rules that must be followed, are specified in the assignment, which is also uploaded. This code was entirely written by me to solve the given problem. The PDF file with the description of the results was created in collaboration with a classmate.